simplifies fixed point lib usage drops unused cast

+++ src/Fractionalizer.sol
@@ -111,9 +111,7 @@ contract Fractionalizer is UUPSUpgradeable, OwnableUpgradeable, ReentrancyGuardU
 
         fractionalized[fractionHash] = fractionalizedToken;
 
-        emit FractionsCreated(
-            uint256(fractionHash), ipnftId, address(fractionalizedToken), _msgSender(), fractionsAmount, agreementCid, name, ipnftSymbol
-        );
+        emit FractionsCreated(fractionHash, ipnftId, address(fractionalizedToken), _msgSender(), fractionsAmount, agreementCid, name, ipnftSymbol);
 
         //if we want to take a protocol fee, this might be a good point of doing so.
         fractionalizedToken.issue(_msgSender(), fractionsAmount);
